\docType{methods}
\name{addsigText.mc}
\alias{addsigText.mc}
\title{Adding significant bar}
\usage{
  addsigText.mc(x0, y0, x1, y1, lab = "", offset)
}
\arguments{
  \item{x0,y0}{coordinates of points from which to draw}

  \item{x1,y1}{coordinates of points to which to draw}

  \item{lab}{label to be added}
}
\description{
  Adding a significant bar to the barplot
}
\author{
  Hoai Tuong Nguyen
}

