\docType{methods}
\name{artif.data.getBE.mc}
\alias{artif.data.getBE.mc}
\title{Calculate Bayes error from a vector of means}
\usage{
artif.data.getBE.mc(mus)
}
\arguments{
  \item{mus}{A vector of means for separation (one class
  would have mus, the other -mus)}
}
\description{
Calculate Bayes error obtained when using vector of means
delta*mu to build artificial data (assuming variance=1)
}
\author{
David Dernoncourt
}

