\docType{methods}
\name{as.numeric.mat.mc}
\alias{as.numeric.mat.mc}
\title{Convert matrix to numeric matrix}
\usage{
  as.numeric.mat.mc(data, dec = ",")
}
\arguments{
  \item{data}{dataframe}

  \item{dec}{decimal of number}
}
\description{
  Convert matrix to numeric matrix
}
\author{
  Hoai Tuong Nguyen
}

