\docType{methods}
\name{barplotStar.mc}
\alias{barplotStar.mc}
\title{Bar Plots with significant bar}
\usage{
  barplotStar.mc(df, title = " ",
    col = c("darkblue", "red"), xlab = "", ylab = "",
    labels)
}
\arguments{
  \item{df}{matrix of values of columns}

  \item{title}{title of bar plot}

  \item{col}{color of two columns}

  \item{xlab}{a label for the x axis}

  \item{ylab}{a label for the y axis}

  \item{labels}{series of columns' names}
}
\description{
  Creates a bar plot with vertical or horizontal bars and
  significant stars.
}
\author{
  Hoai Tuong Nguyen
}

