\docType{methods}
\name{bn.bst.mc}
\alias{bn.bst.mc}
\title{Learning Bayesian networks with bootstraping}
\usage{
bn.bst.mc(data, n)
}
\arguments{
  \item{data}{dataframe}

  \item{n}{number of sample}
}
\description{
Learning Bayesian networks with bootstraping
}
\author{
Hoai Tuong Nguyen
}

