\docType{methods}
\name{cause.mc}
\alias{cause.mc}
\title{Causality analysis}
\usage{
cause.mc(x0, xn, y0, yn)
}
\arguments{
  \item{x0}{values of variable at x0}

  \item{xn}{values of variable at xn}

  \item{y0}{values of variable at y0}

  \item{yn}{values of variable at yn}
}
\description{
Causality analysis
}
\author{
Hoai Tuong Nguyen
}

