\docType{methods}
\name{check.installed.mc}
\alias{check.installed.mc}
\title{Checking package installation}
\usage{
check.installed.mc(pkg)
}
\arguments{
  \item{pkg}{name of package}
}
\value{
A logical value indicating whether the package is installed
}
\description{
Check whether a packages is installed
}
\author{
Hoai Tuong Nguyen
}
\seealso{
\code{\link[utils]{install.packages}}
}

