% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_commit.R
\name{check_commit}
\alias{check_commit}
\alias{cc}
\title{Check Recursively Directories For Uncommitted}
\usage{
check_commit(
  root_directory = ".",
  user = NULL,
  group = NULL,
  commit = FALSE,
  push = FALSE
)

cc(
  root_directory = ".",
  user = NULL,
  group = NULL,
  commit = FALSE,
  push = FALSE
)
}
\arguments{
\item{root_directory}{A directory in which there are git subdirectories.}

\item{group}{A (vector of) group(s) to keep or exclude with prefix "!"..}

\item{commit}{A boolean. if \code{TRUE}, commits uncommitted files.}

\item{push}{A boolean. if \code{TRUE}, push commits.}
}
\value{
A tibble
}
\description{
Check Recursively Directories For Uncommitted
}
