\docType{methods}
\name{FunNet.mc}
\alias{FunNet.mc}
\title{FunNet adapted version}
\usage{
FunNet.mc(wd = "", org = "hsa", two.lists = TRUE, up.frame = NULL,
  down.frame = NULL, genes.frame = NULL, restrict = FALSE,
  ref.list = NULL, logged = FALSE, discriminant = FALSE, go.bp = TRUE,
  go.cc = TRUE, go.mf = TRUE, kegg = TRUE, annot.method = "specificity",
  annot.details = TRUE, direct = FALSE, enriched = TRUE, fdr = NA,
  build.annot.net = TRUE, coexp.matrix = NULL, coexp.method = "spearman",
  estimate.th = FALSE, hard.th = NA, soft.th = NA, topological = FALSE,
  keep.sign = FALSE, level = NA, annot.clust.method = "umilds",
  annot.prox.measure = "unilat.pond.norm.mean", test.recovery = FALSE,
  test.robust = FALSE, replace.annot = NA, random.annot = FALSE,
  build.gene.net = FALSE, gene.clust.method = "hclust",
  gene.net.details = FALSE, gene.clusters = NA, alpha = 0.05, RV = 0.9,
  sigma = NA, keep.rdata = FALSE, zip = TRUE, fileprefix)
}
\description{
FunNet adapted version with customized output filenames
}
\author{
Hoai Tuong Nguyen
}

