\docType{methods}
\name{get.nona.mc}
\alias{get.nona.mc}
\title{Filtering missing value genes}
\usage{
get.nona.mc(df, rate)
}
\arguments{
  \item{df}{a microarray dataframe}

  \item{rate}{percentage of acceptable non-missing values
  per gene}
}
\description{
Filtering missing value genes
}
\author{
Hoai Tuong Nguyen
}

