\docType{methods}
\name{getIgraph.mc}
\alias{getIgraph.mc}
\title{Getting igraph object}
\usage{
getIgraph.mc(nodes, edges, ncolors, ecolors, eweights, directed = TRUE,
  layout = layout.kamada.kawai)
}
\arguments{
  \item{nodes}{list of nodes}

  \item{edges}{list of edges}

  \item{ncolors}{colors of nodes}

  \item{ecolors}{colors of edges}

  \item{eweights}{weights of edges}
}
\description{
Getting igraph object
}
\author{
Hoai Tuong Nguyen
}

