% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mc.R
\name{ggtheme}
\alias{theme_mc}
\alias{theme_mc_md}
\title{Complete MC themes}
\usage{
theme_mc(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

theme_mc_md(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
These are complete themes with or without markdown (\emph{i.e.}, \code{ggtext}).
}
\details{
\describe{

\item{\code{theme_mc()}}{
Theme with \code{c("#333333", "#7F7F7F", "#FAFAFA")} palette from background to foreground.}

\item{\code{theme_mc_md()}}{\code{theme_mc()} with markdown ability using \code{ggtext} (if installed).}

}
}
