\docType{methods}
\name{heatmap.mc}
\alias{heatmap.mc}
\title{Draw a Heat Map}
\usage{
  heatmap.mc(df, col)
}
\arguments{
  \item{df}{matrix of values of columns}

  \item{col}{color of highest value}
}
\description{
  Creates a Heat Map from a matrix of continous values
}
\author{
  Hoai Tuong Nguyen
}

