\docType{methods}
\name{igraph2gephi.mc}
\alias{igraph2gephi.mc}
\title{Convertin igraph object to Gephi object}
\usage{
  igraph2gephi.mc(g, filepath = "converted_graph.gexf")
}
\arguments{
  \item{g}{igraph object}

  \item{filepath}{output file}
}
\value{
  Gephi object
}
\description{
  Convertin igraph object to Gephi object
}
\author{
  Hoai Tuong Nguyen
}

