% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_mom.R
\name{mc_mom}
\alias{mc_mom}
\title{Misclassification correction via mothod of moments}
\usage{
mc_mom(n_jd_P, y_sum_jd_P, n_ji_V, y_sum_ji_V = NULL)
}
\arguments{
\item{n_jd_P}{length-m vector of counts in the primary data by misclassified
groups.}

\item{y_sum_jd_P}{length-m vector of sum(y) in the primary data by
misclassified groups.}

\item{n_ji_V}{a `m`x`m` matrix of counts in the validation data with rows
for misclassified and columns for true groups.}

\item{y_sum_ji_V}{a `m`x`m`  matrix of sum(y) in the validation data with
rows for misclassified and columns for true groups. If NULL, the
`validation` and `with_y_V` estimates will be NaN.}
}
\value{
matrix with the following estimates of the per group mean:
        naive: the naive estimates using the misclassified groups.
        validation: estimates using just the validation data.
        no_y_V: estiamtes with both datasets but without using the `y` value
            in the validation data.
        with_y_V: estimates using all data, including the y value in the
            validation data.
}
\description{
Misclassification correction via mothod of moments
}
\section{Notation}{

See package level docstring for notation.
We use `m` for the number of subgroups.
}

\section{See paper}{

   Selén, Jan. “Adjusting for Errors in Classification and Measurement in
   the Analysis of Partly and Purely Categorical Data.” Journal of the
   American Statistical Association, vol. 81, no. 393, 1986, pp. 75–81.
   JSTOR, www.jstor.org/stable/2287969. Accessed 10 Aug. 2020.
}
