% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rmle.R
\name{mc_rmle}
\alias{mc_rmle}
\title{Misclassification correction via rmle}
\usage{
mc_rmle(n_jd_P, y_sum_jd_P, n_ji_V, y_sum_ji_V, y2_sum_ji_V)
}
\arguments{
\item{n_jd_P}{length-m vector of counts in the primary data by misclassified
groups.}

\item{y_sum_jd_P}{length-m vector of sum(y) in the primary data by
misclassified groups.}

\item{n_ji_V}{a `m`x`m` matrix of counts in the validation data with rows
for misclassified and columns for true groups.}

\item{y_sum_ji_V}{a `m`x`m`  matrix of sum(y) in the validation data with
rows for misclassified and columns for true groups.}

\item{y2_sum_ji_V}{a `m`x`m` matrix of sum(y ^ 2) in the validation data
with rows for misclassified and columns for true groups.}
}
\value{
Matrix with the follow columns:
        mak_li: estimates of mu's.
        variance: estiamtes of variances.
}
\description{
Misclassification correction via rmle
}
\section{Notation}{

See package level docstring for notation.
We use `m` for the number of subgroups.



See package level docstring for notation.
We use `m` for the number of subgroups.
}

\section{See paper}{

T. K. MAK, W. K. LI, A new method for estimating subgroup means under
misclassification, Biometrika, Volume 75, Issue 1, March 1988,
Pages 105–111, https://doi.org/10.1093/biomet/75.1.105
}
