\docType{methods}
\name{samt.mc}
\alias{samt.mc}
\title{Adapted version of samr}
\usage{
  samt.mc(df, label, nperms = 100, logged2 = T,
    type = "Two class unpaired", seq = F)
}
\arguments{
  \item{df}{a microarray dataframe}

  \item{label}{vector of label}

  \item{nperms}{number of permutations}

  \item{logged2}{logical. Expression level has been
  transformed by logorith base 2}
}
\description{
  An adapted version of samr for gene differtiation
  analysis
}
\author{
  Hoai Tuong Nguyen
}

