\docType{methods}
\name{signs.mc}
\alias{signs.mc}
\title{Get signs of changes}
\usage{
  signs.mc(t0, tn)
}
\arguments{
  \item{t0}{values of variable at t0}

  \item{tn}{values of variable at tn}
}
\description{
  Get signs of changes for a variable at two timepoints
}
\author{
  Hoai Tuong Nguyen
}

