\docType{methods}
\name{updateKDB.mc}
\alias{updateKDB.mc}
\title{Updating knowledge database}
\usage{
  updateKDB.mc(newKDB.or, nodes = NULL)
}
\arguments{
  \item{arcs}{list of edges}

  \item{method}{method used to get edges}

  \item{param}{parameters of method}

  \item{dbfiles}{path storing data files}
}
\description{
  Updating knowledge database
}
\author{
  Hoai Tuong Nguyen
}

