\docType{methods}
\name{weighted.var.mc}
\alias{weighted.var.mc}
\title{Calculate weighted variance}
\usage{
weighted.var.mc(x, w = NA, na.rm = FALSE)
}
\arguments{
  \item{x}{vector of values}

  \item{w}{weights of each values}

  \item{na.rm}{boolean, whether or not remove NAs}
}
\description{
Calculate weighted variance, original function posted by
Dr. Gavin Simpson:
https://stat.ethz.ch/pipermail/r-help/2008-July/168762.html
}
\author{
David Dernoncourt
}

