\docType{methods}
\name{yule.Q.mc}
\alias{yule.Q.mc}
\title{Correlation between nominal variables}
\usage{
yule.Q.mc(x, y)
}
\arguments{
  \item{x}{a nominal vector}

  \item{y}{a nominal vector}
}
\description{
A measure of correlation between nominal variables
}
\author{
Hoai Tuong Nguyen
}

