\docType{methods}
\name{boxplot.class.mc}
\alias{boxplot.class.mc}
\title{Box Plots}
\usage{
boxplot.class.mc(data, x, type = "auto", class, xlab, ylab, outfile = NULL)
}
\arguments{
  \item{data}{a data frame}

  \item{x}{index of column}

  \item{class}{vector of classes}

  \item{outfile}{output filename}
}
\description{
Draw a boxplot for a column of data frame
}
\author{
Hoai Tuong Nguyen
}

