\name{MC-package}
\alias{MC-package}
\alias{MC}
\docType{package}
\title{
\packageTitle{MC}
}
\description{
\packageDescription{MC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MC}
\packageIndices{MC}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MC}

Maintainer: \packageMaintainer{MC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
