% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc-package.R
\docType{package}
\name{mc-package}
\alias{mc}
\alias{mc-package}
\title{mc: Mickaël Canouil}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Bunch of functions mostly for my personal use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcanouil/mc}
  \item Report bugs at \url{https://github.com/mcanouil/mc/issues}
}

}
\author{
\strong{Maintainer}: Mickaël Canouil \email{mickael.canouil@cnrs.fr} (\href{https://orcid.org/0000-0002-3396-454}{ORCID})

}
\keyword{internal}
