% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_sig.R
\name{mc_sig}
\alias{mc_sig}
\title{Return the standard New Graph email signature as HTML}
\usage{
mc_sig()
}
\value{
A character string of HTML.
}
\description{
Reads the signature template bundled at \code{inst/sig/signature.html}.
Used by \code{\link[=mc_md_render]{mc_md_render()}} to append the signature automatically.
}
\examples{
\dontrun{
cat(mc_sig())
}

}
